Convert R objects to/from strings

Serialise/deserialise an R object into a string. This is a very thin wrapper around the existing R functions serialize() and rawToChar(). This is useful to encode arbitrary R objects as string to then save in Redis (which expects a string).

Usage

object_to_string(obj)

string_to_object(str)

object_to_bin(obj, xdr = FALSE)

bin_to_object(bin)

Arguments

obj

An R object to convert into a string

str

A string to convert into an R object

xdr

Use the big-endian representation? Unlike, serialize() this is disabled here by default as it is a bit faster (~ 20%, saving about 1 microsecond of a 5 microsecond roundtrip for a serialization of 100 doubles)

bin

A binary vector to convert back to an R object

Examples

s <- object_to_string(1:10)
s
#> [1] "A\n3\n263425\n197888\n5\nUTF-8\n238\n2\n1\n262153\n14\ncompact_intseq\n2\n1\n262153\n4\nbase\n2\n13\n1\n13\n254\n14\n3\n0x1.4p+3\n0x1p+0\n0x1p+0\n254\n"
string_to_object(s)
#>  [1]  1  2  3  4  5  6  7  8  9 10
identical(string_to_object(s), 1:10)
#> [1] TRUE