Load Lua scripts into Redis

Load Lua scripts into Redis, providing a convenience function to call them with. Using this function means that scripts will be available to use via EVALSHA, and will be preloaded on the Redis server. Scripts are then accessed by name rather than by content or SHA. See vignette("redux") for details and an example.

Usage

redis_scripts(con, ..., scripts = list(...))

Arguments

con

A Redis connection

...

A number of scripts

scripts

Alternatively, a list of scripts